Generate fixtures script
